started working on battle scene refactor
