+++ src/scenes/battle-scene.js
@@ -7,18 +7,6 @@ import {
 import Phaser from '../lib/phaser.js';
 import { SCENE_KEYS } from './scene-keys.js';
 
-const BATTLE_MENU_OPTIONS = Object.freeze({
-  FIGHT: 'FIGHT',
-  SWITCH: 'SWITCH',
-  ITEM: 'ITEM',
-  FLEE: 'FLEE',
-});
-
-const battleUiTextStyle = {
-  color: 'black',
-  fontSize: '30px',
-};
-
 export class BattleScene extends Phaser.Scene {
   constructor() {
     super({
@@ -131,32 +119,5 @@ export class BattleScene extends Phaser.Scene {
       .setScale(1, scaleY);
 
     return this.add.container(x, y, [leftCap, middle, rightCap]);
-  }
-
-  #createMainInfoPane() {
-    const padding = 4;
-    const rectHeight = 124;
-
-    this.add
-      .rectangle(
-        padding,
-        this.scale.height - rectHeight - padding,
-        this.scale.width - padding * 2,
-        rectHeight,
-        0xede4f3,
-        1
-      )
-      .setOrigin(0)
-      .setStrokeStyle(8, 0xe4434a, 1);
-  }
-
-  #createMainInfoSubPane() {
-    const rectWidth = 500;
-    const rectHeight = 124;
-
-    return this.add
-      .rectangle(0, 0, rectWidth, rectHeight, 0xede4f3, 1)
-      .setOrigin(0)
-      .setStrokeStyle(8, 0x905ac2, 1);
   }
 }
